refactor directory organization
