+++ components/details/Ingredient.module.css
