+++ components/shared/Footer.js
@@ -1,7 +1,7 @@
 import Link from "next/link";
 
-import HouseIcon from "../icons/HouseIcon";
-import CartIcon from "../icons/CartIcon";
+import HouseIcon from "../../icons/HouseIcon";
+import CartIcon from "../../icons/CartIcon";
 
 import styles from "./Footer.module.css";
