+++ components/shared/Header.js
@@ -2,7 +2,7 @@ import Image from "next/image";
 
 import styles from "./Header.module.css";
 
-import logo from "../public/images/logo2.png";
+import logo from "../../public/images/logo2.png";
 
 export default function Header() {
   return (
