+++ components/shared/TrashButton.js
