+++ lib/createSlices.js
