+++ lib/deepCopy.js
