+++ pages/details/details.module.css
